refactor: move model interface to correct place
