+++ internal/usecase/interfaces/interfaces.go
